Added model manager, start-at and end-at fields.
The model manager provides functionality to find upcoming live chats to
advertise on the home page and askmama page. It also provides a function to
return the currently in progress live chat, to allow redirecting the user to the
live chat directly from the AskMAMA link in the top navigation, and the
advertisement.